requirements.txt
asgiref==3.3.4
Cython==0.29.23
Django==3.2.4
djangorestframework==3.12.4
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
